feat: change project structure
