+++ lib/src/enums.dart
@@ -1,0 +1,7 @@
+enum TooltipDirection { up, down, left, right }
+
+enum ShowCloseButton { inside, outside, none }
+
+enum ClipAreaShape { oval, rectangle }
+
+enum Event { show, hide }
